refactor(backend): extract direct stream text helpers
- Move direct visible thinking and answer stream text helpers into direct_stream_text_runtime.py
- Keep direct_execution as provider invocation and fallback shell with compatibility re-exports for graph shims
- Update architecture and cleanup audit docs with verification, risk, and rollback notes
